pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Evgeniy Slizh"]
readme = "README.md"
packages = [{include = "task_manager"}]


[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django = "^5.1.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
uvicorn = "^0.31.0"
django-bootstrap5 = "^24.3"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
django-filter = "^24.3"
rollbar = "^0.16.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
